task_tracker/pyproject.toml
[tool.poetry]
name = "task_tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
motor = "^3.0.0"
fastapi = "^0.85.0"
uvicorn = {extras = ["standart"], version = "^0.18.3"}
pydantic = "^1.10.2"
nats-py = "^2.2.0"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
